+++ packages/suite/src/actions/suite/constants/metadataConstants.ts
@@ -1,6 +1,7 @@
 export const ENABLE = '@metadata/enable';
 export const DISABLE = '@metadata/disable';
 export const SET_DEVICE_METADATA = '@metadata/set-device-metadata';
+export const SET_DEVICE_METADATA_PASSWORDS = '@metadata/set-device-metadata-passwords';
 export const ADD_PROVIDER = '@metadata/add-provider';
 export const REMOVE_PROVIDER = '@metadata/remove-provider';
 export const ACCOUNT_ADD = '@metadata/account-add';