+++ packages/suite/src/actions/suite/metadataActions.ts
@@ -15,7 +15,7 @@ import {
 import * as metadataUtils from 'src/utils/suite/metadata';
 import { selectSelectedProviderForLabels } from 'src/reducers/suite/metadataReducer';
 import { Account } from '@suite-common/wallet-types';
-import type { AbstractMetadataProvider } from 'src/types/suite/metadata';
+import type { AbstractMetadataProvider, PasswordManagerState } from 'src/types/suite/metadata';
 
 export type MetadataAction =
     | { type: typeof METADATA.ENABLE }
@@ -24,6 +24,10 @@ export type MetadataAction =
     | { type: typeof METADATA.SET_INITIATING; payload: boolean }
     | {
           type: typeof METADATA.SET_DEVICE_METADATA;
+          payload: { deviceState: string; metadata: DeviceMetadata };
+      }
+    | {
+          type: typeof METADATA.SET_DEVICE_METADATA_PASSWORDS;
           payload: { deviceState: string; metadata: DeviceMetadata };
       }
     | {
@@ -125,7 +129,7 @@ export const setMetadata =
     }: {
         provider: MetadataProvider;
         fileName: string;
-        data: WalletLabels | AccountLabels | undefined;
+        data: WalletLabels | AccountLabels | PasswordManagerState | undefined;
     }) =>
     (dispatch: Dispatch) => {
         dispatch({
@@ -145,7 +149,7 @@ export const encryptAndSaveMetadata = async ({
     fileName,
     providerInstance,
 }: {
-    data: AccountLabels | WalletLabels;
+    data: AccountLabels | WalletLabels | PasswordManagerState;
     aesKey: string;
     fileName: string;
     providerInstance: AbstractMetadataProvider;