+++ packages/suite/src/middlewares/suite/suiteMiddleware.ts
@@ -38,6 +38,7 @@ const isActionDeviceRelated = (action: AnyAction): boolean => {
     }
 
     if (action.type === METADATA.SET_DEVICE_METADATA) return true;
+    if (action.type === METADATA.SET_DEVICE_METADATA_PASSWORDS) return true;
 
     if (Object.values(DEVICE).includes(action.type)) return true;