+++ packages/suite/src/reducers/suite/__fixtures__/deviceReducer.ts
@@ -905,6 +905,7 @@ const changePassphraseMode: Fixture<ReturnType<typeof deviceActions.updatePassph
                         ts: 1,
                         buttonRequests: [],
                         metadata: {},
+                        passwords: {},
                     },
                     hidden: false,
                 },