+++ packages/suite/src/utils/suite/passwords.ts
@@ -1,0 +1,13 @@
+export const getDisplayKey = (title: string, username: string) => {
+    try {
+        return `Unlock ${new URL(title).host} for user ${username}?`;
+    } catch (err) {
+        return `Unlock ${title} for user ${username}?`;
+    }
+};
+
+export const getNextId = (entries: Record<number, any>) => {
+    if (Object.keys(entries).length === 0) return 0;
+
+    return Number(Object.keys(entries).sort((a, b) => parseInt(b, 10) - parseInt(a, 10))[0]) + 1;
+};